- fixed several warnings when compiling with -Wall flag (Alexei)
git-svn-id: svn://svn.zabbix.com/trunk@2222 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/libs/zbxcommon/str.c
@@ -1,6 +1,7 @@
 #include "common.h"
 
 #include <string.h>
+#include <stdlib.h>
 
 /* Has to be rewritten to avoi malloc */
 char *string_replace(char *str, const char *sub_str1, const char *sub_str2)
@@ -27,7 +28,7 @@ char *string_replace(char *str, const char *sub_str1, const char *sub_str2)
         diff = strlen(sub_str2) - len;
 
         /* allocate new memory */
-        if ( (new_str=(char *)malloc((strlen(str) + count*diff)*sizeof(char)))
+        if ( (new_str=(char *)malloc((size_t)(strlen(str) + count*diff)*sizeof(char)))
                         == NULL )
                 return NULL;
